chore: instrument experiment
